[roll third_party/pigweed] pw_metric: add _STATIC variants
This adds PW_METRIC_STATIC and PW_METRIC_GROUP_STATIC, which do the same
as their non-_STATIC siblings, except that the object that they create
gets a `static` storage.

This can be useful, e.g. to have metrics inside of functions, when the
function itself is to be instrumented.

This CL also removes an extraneous stringification in the metric and
group names, which resulted in double double quotes.

Ah, and I also implemented PW_METRIC_GROUP with 3 parameters, which was
referenced in the documentation but not formally implemented.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/16980

Rolled-Commits: 7a5e4d66bd94923..2182c6688ca3173
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I08e9912a0116bd1d2cdec17489b33573bd7c5676
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/17163
Reviewed-by: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>