[roll third_party/pigweed] pw_rpc: Send CLIENT_ERROR on unexpected packet
This updates the RPC client to send back a CLIENT_ERROR to the server if
it receives a packet it was not expecting. The server is updated to
handle client errors by ending the active RPC, if applicable.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/19180

Rolled-Commits: da651f169f97f85..062ed18abbf7d25
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I1d64f50ef4da31c3d222a28b85271506eaae1b8d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/19600
Reviewed-by: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>