[roll third_party/pigweed] pw_rpc: Method union class
This adds a MethodUnion class which can store one of multiple
method types, and updates the base Service class to store them instead
of concrete Methods. A MethodUnion class for raw methods and for
nanopb/raw methods is added to the respective submodules. Also, the
nanopb codegen API and script are updated to deduce a method type from
an implementation's signature, and create the invoker automatically.

This does not update the TestMethodContext class to work with the new
changes; tests depending on it are temporarily disabled until fixed.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/21242

Rolled-Commits: e0575f7256215b2..a4d7150769f87ff
CQ-Do-Not-Cancel-Tryjobs: true
Roller-URL: https://ci.chromium.org/b/8865726975114069168
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8865726975114069168
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I2d69fdf0d35f8a9e994eebde05aa8c1c9ab21bf0
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/22185
Reviewed-by: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>