[roll third_party/pigweed] pw_trace: improve locks
Add an internal queue with a separate lock, this is a small section
which just copies the arguments making it suitable for a critical
section.

Emptying this queue is handled from a separate lock, which is a trylock.
If the queue is already being emptied, it is safe to return after adding
to the queue (no need to block).

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/21881

third_party/pigweed Rolled-Commits: 5a3a61cd3bf16da..5077c7c365705a2
CQ-Do-Not-Cancel-Tryjobs: true
Roller-URL: https://ci.chromium.org/b/8865173881552864432
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8865173881552864432
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5d830eb3919c082ef7a94e95054920948d9e7a30
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/22706
Reviewed-by: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>