[roll third_party/pigweed] pw_presubmit: Adds Bazel support to presubmit
This adds initial support for Bazel presubmit based on the modules
that are known to be working. The intention here is that after Bazel
support is completed for each module it is added to the list of
modules to be built and tested. This should prevent any build system
regressions.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/39101

third_party/pigweed Rolled-Commits: df38ed19386233a..40ab6b0d57e3dd1
Roller-URL: https://ci.chromium.org/b/8851320184344582192
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8851320184344582192
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ia28b2dcdd5b10df77a08a21cb7c7759666df3a0f
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/39561
Reviewed-by: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>