[roll third_party/pigweed] pw_ring_buffer: Add late-reader/drain attach
Changes the default behavior of the ring buffer and multisink such that
late-attach of a Reader or Drain now places the object at the index of
the oldest entry in the buffer.

For Readers, the oldest entry corresponds to the slowest reader present.
If no readers are present, the Reader is placed at the write head. For
Drains, a lagging Reader is always present, so new Drains will always
point to the oldest valid entry in the buffer.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/53344

third_party/pigweed Rolled-Commits: 3eb97d4c0673aa1..099f716e17f2bb3
Roller-URL: https://ci.chromium.org/b/8841583512165953008
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8841583512165953008
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I9a41ee2af78a3cf5358a66aee3731f8b46c6077b
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/53460
Reviewed-by: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>