[roll third_party/pigweed] pw_transfer: Refactor protocol code
This moves most of the transfer logic into a common context class that
can be shared between and client and a server. The transfer service is
updated to use this class.

This class differs from the original transfer implementation in that is
defines its operations in two parts: an immediate action and a deferred
action. This is done to avoid blocking an RPC thread for too long with
potentially long stream reads/writes. As part of this, the transfer
service now requires a memory buffer in which data can be staged for
later use.

Original-Requires: pigweed-internal:17300
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/67032

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 59910a47b1d091d..d71f4f66c3303cc
Roller-URL: https://ci.chromium.org/b/8831012996311207009
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8831012996311207009
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ia34dfec9c90be834184bb0cab1e2a0c40b5c3235
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/68042
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>