[roll third_party/pigweed] pw_build: Refactor pw_mirror_tree with copy
The pw_mirror_tree GN template is used to "copy" a set of files from one
directory to another, using a python script that created hard links from
the source files to the output files. This led to a subtle issue where
it is possible for incremental builds to evaluate incorrectly and
execute actions that are not necessary.

Consider the following ninja build rule:

  build output_file : pw_mirror_tree_rule input_file | pw_mirror_tree.py

When ninja is considering whether to rebuild output_file, it will
compare the output_file's mtime to the most recent mtime of all of the
inputs. Due to the use of hard links, the mtime of output_file will be
the same as input_file, so this means ninja will be comparing the
mtimes of input_file and pw_mirror_tree.py.

This CL refactors the pw_mirror_tree template in terms of GN copy rule
instead of GN actions. The GN copy rule expects to be implemented using
hard links, so it ensures that all dependencies are treated as
order-only and mtimes will not be evaluated.

Original-Bug: 554
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/68902

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 74b888156b98d13..13845149c264a3f
Roller-URL: https://ci.chromium.org/b/8830452845904322481
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8830452845904322481
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Icb79270ed7dc1b769754da2c8550e4015587febc
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/69102
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>