[roll third_party/pigweed] pw_software_update/py: Relax root metadata signing checks
When signing a root metadata, the signing key does not need to be listed
in the root metadata. This CL removes those checks.

Use case: When rotating a root key, the new root metadata need to signed
with both the existing root key and the new root key but the old root
key does NOT need to be listed in the new root metadata. In fact if the
rotation aims to revoke the old root key, the old root key MUST NOT be
listed in the new root signature requirements.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/69360

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: e2e23280635baad..4a555a4db0f659b
Roller-URL: https://ci.chromium.org/b/8829157558635140193
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8829157558635140193
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0d76aec39ba66f0abd6c3a45186ec9c00d37df51
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/71302
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>