[roll third_party/pigweed] pw_protobuf: Support providing known length to StreamDecoder
When a protobuf is encoded on a stream, the wire format may include
advance details of the length of the message.

Allow this to be specified to StreamDecoder so that it does not read
(and consume) bytes beyond the bounds of the message to be decoded.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/74722

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: d8ac6bdc1d04fc4..27e12c5bc7dd788
Roller-URL: https://ci.chromium.org/b/8825929184851326641
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8825929184851326641
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0544b2453f5aa7777507f2feb9f0f19b1389da46
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/76720
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>