[roll third_party/pigweed] pw_cpu_exception_cortex_m: Expose LogExceptionAnalysis()
Exposes an internal part of the support backend to log the
exception analysis as an optional utility API.

The cpu_state.cc file is renamed to support.cc to reflect what
the facade/backend that the file is associated with.

The cpu_state.h is pulled out to a separate target to remove
unnecessary circular dependencies within the module.

In addition the last remaining TODOs for pwbug/296 are removed and
the Bazel build files are cleaned up a little bit.

Also updates the docs to include the configuration options.

Original-Bug: 296
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/78860

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 6eafce383bcce53..49f21314ece3a23
Roller-URL: https://ci.chromium.org/b/8825020880827391057
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8825020880827391057
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ib333cb5fd16977a8ad387b2ba47aab8d749a9645
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/78920
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>