[roll third_party/pigweed] pw_protobuf: Correct MaxScratchBufferSize type
It is unusual for a free function in a header to be marked static,
because it tells the compiler to generate a separate function in each
translation unit that includes it
(https://stackoverflow.com/a/22103109). It makes more sense to mark it
for inlining.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/79260

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: a0fc38ce53e80a2..6e92d78fe513102
Roller-URL: https://ci.chromium.org/b/8824590041888124161
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8824590041888124161
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I4f475700ac5fece6bec680093110dce2701c9261
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/79420
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>