[roll third_party/pigweed] third_party/freertos: Always provide pw_third_party.freertos
Updates the CMake build to always provide the
pw_third_party.freertos target similar to how facades work. Unless
configured it is an interface library which depends on an actionable
error command which instructs the user if it is built that it must be
configured first.

This ensures that the CMake build can evaluate the target regardless
of the toolchain and configuration that is being used.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/82760

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: d0cca897917ab3e..4404eb53ea6cdec
Roller-URL: https://ci.chromium.org/b/8823117978319256145
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8823117978319256145
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I35d82fd27c0abfe5fe7aaaf3921cfd55a18c80e7
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/83075
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>