[roll third_party/pigweed] pw_transfer: Protocol versioning and chunk refactor
This adds the concept of versioning to the C++ pw_transfer
implementation. The version of a given transfer is deduced from the
chunks sent and received based on the fields which are present within
them. Notably, there is no explicit version number encoded anywhere on
the wire.

The addition of versioning is done in such a way that transfer
implementations can remain backwards compatible, supporting
communication with users of older protocol versions.

Initially, only a single "legacy" version is supported and used, with
its handling behavior matching the current uses of pw_transfer. An
improved protocol (tentatively "v2") will be rolled out over a series
of subsequent CLs.

To support these goals, the transfer chunk and code using it is
refactored into an encapsulating class, with internal logic for encoding
and decoding different fields based on its configured or inferred
version. A common public API is presented to the rest of the transfer
code, allowing it to be written in a (mostly) version-agnostic manner.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/92640

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 56e055a1a2bb4bc..42efd503284729b
Roller-URL: https://ci.chromium.org/b/8814508137567719697
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8814508137567719697
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5a7cd47f2cde4e9112bed99c13b65e66f9ab43ee
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/94145
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>