[roll third_party/pigweed] pw_toolchain: Fix clang-tidy.py multi-arg support
There is was an assumption that after the -- in the clang-tidy
command there would be exactly one word and that would be the
clang binary name.  This prevented having anything else in the
string, including things necessary for RBE.  This change,
adds a marker 'END_OF_INVOKER' to the end of the compile
argument so it can explicitly be checked for so additional
arguments can be added before or after the compile binary.

Test: Verified this generated the build files correct and built
      correctly with RBE and regular builds.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/98360

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 3f78d8202ba7575..2ba0413a1ffeafe
Roller-URL: https://ci.chromium.org/b/8810088585517510801
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8810088585517510801
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I3cbc9aacfc787636ef9429bd1d1f87beb969759b
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/100281
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>