[roll third_party/pigweed] pw_sync: Clarify PW_LOCKS_EXCLUDED documentation
The documentation of PW_LOCKS_EXCLUDED previously stated that the
annotated function acquired the lock in question. However, this is only
one possible use of the annotation. The annotated function may wish to
guarantee that a lock is not held for other purposes, such as ensuring
that the code in question is not run as part of a critical section.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/103421

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 41af1d856124d12..0117408af4b6c33
Roller-URL: https://ci.chromium.org/b/8807620959259703025
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8807620959259703025
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I4ffcdfb41f422e436653da1147e778caec5a93de
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/103501
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>