[roll third_party/pigweed] pw_unit_test: Export missing types from Python module
The `EventHandler` type, which is exported from the pw_unit_test Python
module, has functions to override that take the TestExpectation
type as arguments. However, the TestExpectation type is not exposed in
the module's __init__.py (unlike other related types), so it's not
easily accessible from dependent code. This fixes the issue.

Additionally, EventHandler has a function that takes the result of a
test case. This is passed as an `int`, using the value of the generated
internal protobuf. To use this in dependent code, you'd need to either
find and use the generated protobuf (which isn't clear because the type
isn't listed), or use magic constants in the code. This commit also
addresses this issue by wrapping these results in an enum
TestCaseResult and exporting that.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/103486

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: bc59ea28cf3c442..f89426038c046b7
Roller-URL: https://ci.chromium.org/b/8807439203895449969
Cq-Cl-Tag: roller-builder:pigweed-sample-project-roller
Cq-Cl-Tag: roller-bid:8807439203895449969
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I4d83f8a7a31a0460d30bcbd91390ce352a9f576d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/103721
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>