[roll third_party/pigweed] pw_transfer: Version 2 opening handshake in C++
This implements the opening handshake of pw_transfer's version 2
protocol within the C++ transfer client and server. The handshake
consists of protocol version negotiation and ephemeral session ID
assignment.

The protocol version used for a transfer session is controlled by the
client when it starts a new transfer. By default, this still remains the
legacy protocol for now, though the client API is extended to allow
specifying the desired version.

If the START chunk a transfer service receives is configured for version
2, the service will assign a transfer session ID and proceed with the
handshake. Otherwise, it will fall back to the legacy protocol.

The version 2 START chunk sent by a client retains all of the chunk
proto fields set by the legacy protocol, allowing it to be understood
by a server which is not version 2 aware. In such a case, the server
will process the chunk per the legacy protocol and send a non-handshake
response. The client will recognize the legacy response and revert to
running the legacy protocol.

As a result of this, version 2 capable transfer clients and servers
remain fully backwards-compatible with older code that only runs the
legacy protocol.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/99500

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 400c0f3bdb19675..64a88a417ce3e29
Roller-URL: https://ci.chromium.org/b/8806003560920685921
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I6a9b53d4bc15a35044db1bcfd94d206b81f500be
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/106321
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>