[roll third_party/pigweed] pw_stream: Implement StdFileReader::ConservativeLimit
The ConservativeLimit method returns an estimate of how much data is
available to be read or written in a pw::stream. The default
implementation returns a very large value ("unlimited").

This commit implements a file-specific implementation for StdFileReader.
It determines the size of the file (and thus how much data is left to
read) by seeking to the end and getting the current file position. This
makes the ConservativeReadLimit method much more useful for users of the
class.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/106130

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 26e210529ec4d39..3bac75d77b9b492
Roller-URL: https://ci.chromium.org/b/8805741365016905489
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5b7d5a3dc390c74f1ef9a194b1eb697c2ccb93c1
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/106630
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>