[roll third_party/pigweed] pw_build: Don't mutate deps list in bazel macros
The deps list is passed by reference. If the caller uses a variable as
the deps list, rather than a literal, that variable will be mutated as
a side effect of the macro invocation if we use `append`. This is not
intended.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/106912

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: ac7f0090500ed5b..79c1756eeb72e81
Roller-URL: https://ci.chromium.org/b/8805450035842229025
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Idde016e1d0c51d697447c179697bd46f159ff528
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/107511
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>