[roll third_party/pigweed] pw_transfer: Use correct chunk ID when resource doesn't exist
In transfer protocol v2, session IDs are assigned by the server. When a
client sends its opening chunk, it does not yet know its session ID and
expects a response that is identified by the resource ID it is trying to
transfer.

In response to an opening chunk from the client, the server looks up a
handler for the requested resource. If one is not found, the transfer
thread responds directly without assigning a context for the transfer.
This direct response did not account for the initial session/resource ID
discrepancy, instead sending its NOT_FOUND chunk with the newly assigned
session ID of which the client had no knowledge.

This updates the transfer thread to always use resource ID to identify
an error in response to an opening chunk. In legacy protocol, the two
are the same, whereas in v2, resource ID is expected.

Original-Bug: 252870476
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/113990

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: c60af251ba2d91f..770673f4cfd4f03
Roller-URL: https://ci.chromium.org/b/8800645794655701057
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I4e6f585597ff46771fbc172f73677164ffa71ee1
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/113974
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>