[roll third_party/pigweed] pw_presubmit: Allow reuse of filtered files
When a presubmit program is run in CQ, it's first run with
'--only-list-steps'. Recipe code parses the results of this, which shows
which steps are to be run based on the files changed. Then the steps are
each run individually. This makes the UI of CQ better, because each step
shows separately. However, we don't need to reevaluate the changed files
for each step.

Change the behavior of '--only-list-steps' to also output the list of
files that should be passed into each step. Then add a
'--list-steps-file' argument that reads in this list of files rather
than recalculating it from the checkout on disk.

Since recipe code does not yet use '--list-steps-file', this change has
no immediate effect except to add a list of paths to the output of
'pw presubmit --only-list-steps', which recipe code will ignore.

Original-Bug: b/256007641
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/116576

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 214d603e06dcbe8..fcff2e904165413
Roller-URL: https://ci.chromium.org/b/8798632986443504337
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I57e2969dd36f9a7065304bb7962873735ed964ba
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/116725
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>