[roll third_party/pigweed] pw_unit_test: Add code coverage awareness
This changes `test_runner.py` to be aware of the intention to run a code
coverage test (any test that is executed with the automatic runner when
`pw_toolchain_SANITIZERS` contains `"coverage"`). This is accomplished
by adding an extra `argparse` arg called `--run-for-coverage`.

When that flag is set, the test runner will set the `LLVM_PROFILE_FILE`
environment variable to generate a `profraw` file next to the test
binary.

The `pw_test` template is also updated to check for `"coverage"` in
`pw_toolchain_SANITIZERS`, and sets `--run-for-coverage` and includes
the `profraw` file in outputs when in coverage mode.

The `profraw` file path for the test is then propagated through
`metadata.profraws` similar to `tests`. This allows a `generated_file`
to be used downstream to collect all `profraw` paths.

Further work is required to use these `profraw` files to actually
generate coverage reports with `llvm-cov` or some other tool, but this
is a first step.

Tested: `pw presubmit`.
Tested: Downstream project continues to build in coverage mode and succesfully outputs `profraw` files.
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/116317

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: b2a670839ede5d1..057172e3c58deed
Roller-URL: https://ci.chromium.org/b/8798572595517641521
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I3efb58eb5400111bf222e2c81bc089f886970a02
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/117032
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>