[roll third_party/pigweed] pw_protobuf: Put codegen in a ::pwpb suffix namespace
The namespace suffix is applied at the package level, as specified in
the originating .proto file.

This change is mostly in order to avoid clashing with code generated by
the stock C++ representation outputted by protoc. This is useful, for
example, for projects that would like to use the stock C++ proto library
for additional tests, but which need to use pw_protobuf due to needs of
embedded targets.

For now, this change is set up in a way so that it is, by default,
backwards compatible with existing code. A flag has been added to the
plugin to opt out of legacy behaviour, but has not been implemented in
any rules that invoke the Pigweed proto compiler plugin.

Because of limitations with the current proto plugin implementation,
some external identifiers (references to things in other .proto files,
typically) are not parsed with a package annotation, and so we are
unable to codegen the ::pwpb suffix in the correct place. For now, I've
introduced an auxiliary namespace (discussed with hepler@google.com,
seems like the best fast workaround).

In the long term, potentially we can coax the plugin to traverse the
build graph better. (Package or namespace overrides are supported for
Java / Obj-C / C#, so presumably some mechanism is available.)

b/250945489

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/117934

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: f7c3a1e61a28d58..f61993aad0077ef
Roller-URL: https://ci.chromium.org/b/8797740151821244481
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I8ff1c9cdd39b8c142fd9b9eb352612a72c625315
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/118928
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>