[roll third_party/pigweed] pw_build: Add pw-wrap-ninja helper
This commit adds a standalone Python script that wraps Ninja to provide
a more comprehensive display of the progress of the build, along with a
brief summary of the tool to the pw_build docs.

The tool works by parsing Ninja's output to build a representation of
the build status, and then outputting a richer TUI with the list of
running targets and how long they're taking. This allows one to, at a
glance, determine what targets are blocking the rest of the build or
taking an unusually long time to complete.

Original-Bug: 263976434
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/125551

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: d1a9833caf1704c..21a65831d7e63b8
Roller-URL: https://ci.chromium.org/b/8793393982443226721
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Iee0a05d75e201610bd9ce9aa9ad8533ed5740e8a
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/125711
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>