[roll third_party/pigweed] pw_build: Make pw-wrap-ninja avoid terminal codes if non-interactive
This commit makes the pw-wrap-ninja script avoid the use of ANSI
terminal codes if it detects that it is running in a non-interactive
context.

This follows the same logic as Ninja: a terminal is interactive if
all of the following are true:
1) stdout is a TTY
2) the TERM environment variable is set
3) the TERM environment variable is not "dumb"

In this mode, the script will avoid printing out the live progress
display (as this requires terminal codes). Instead, it'll just print out
whenever an action finishes, along with any output from the action.

Tested:
  Ran pw-wrap-ninja piped to tee, saw that the progress display
  was hidden, no ANSI escapes were visible, and the 'Finished...' lines
  were printed.
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126172

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 595ea383848eb3a..6af9dbfe17e0ddc
Roller-URL: https://ci.chromium.org/b/8791681712949903649
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ia0059f3dec88d4e987f46b62d2efec77a64f51b3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/126664
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>