[roll third_party/pigweed] pw_tokenizer: Add tests for %% specifier
This makes a decision to adhere to the C99 specification that `%` is a
specifier only when used exactly as `%%`, without any flag, width,
precision, or length modifiers. This was the current implementation
since Python 3 has issues with things like `%+%`, but it is worth noting
that both clang and gcc do allow any modifiers (and simply ignore them).

This can be changed in the future as long as documentation and tests are
updated.

Original-Bug: b/265307572
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/126838

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: cbb713112a69582..202bffdfd8d206a
Roller-URL: https://ci.chromium.org/b/8789680501726236097
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I882185cb4bca4d9eaef3e1b3efd9ca9b6dc2fcf3
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/128604
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>