[roll third_party/pigweed] pw_build: bazel: Make pw_assert dep unconditional
This change makes Pigweed's pw_cc_binary, pw_cc_test, pw_cc_perf_test
Bazel macros _unconditionally_ add a dependency on the //pw_assert
_backend_, instead of _conditionally_ depending on //pw_assert itself.

This is technically a breaking change: any targets of these types which
already explicitly included a direct dependency on the //pw_assert
backend will break. However, this is not expected to break anyone in
practice, because binaries should be directly depending on //pw_assert
itself, not the backend.

The motivation for this change is that the implementation of conditional
dependency addition contained a bug. It produced build errors whenever
the `deps` of a pw_cc_binary included a `select` statement. (See the
issue referenced below for a discussion of the mechanism of the bug.)
This is an important bug, because multi-platform builds naturally lead
to using `select` statements to produce different build graphs when
building for different platforms. Unconditionally adding the //pw_assert
backend to the deps will not suffer from this problem.

Original-Bug: b/234877642
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/129373

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 94d0b3c876fbd9c..ee69427379a8f68
Roller-URL: https://ci.chromium.org/b/8787887226744446417
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I6a7ae5ce26743532bae1e0c830c9e5f97f6a275d
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/131354
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>