[roll third_party/pigweed] pw_i2c: make Device objects movable
Currently the Device object is not copyable or movable. This change
adds back the default move constructor. The Device is just a wrapper
around an address and Initiator reference, so it is safe to move.

Add documentation about the concurrency requirements of Initiator and
Device. Explain that Device is intended to represent ownership, and
that's why it is not copyable.

This change is intended to make it easier to construct objects which
depend on a Device. Previously the Device had to be allocated separately
and passed by reference to the constructor. This change allows using
temporary Device objects when constructing higher level objects.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/134610

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 0ad7f43af24e35f..6e6232b7777cfa7
Roller-URL: https://ci.chromium.org/b/8786004427795137601
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0ad8ee6f6c3d29d9d169ebb134fd429034fd38b9
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/134696
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>