[roll third_party/pigweed] pw_ide: Broaden build system support
In contrast to GN, which has a single top-level build that manages all
targets, CMake generates a separate Ninja build for each target. This is
actually the typical case, and GN's approach is atypical. pw_ide now
takes a more hands-off approach and simply finds compilation databases
regardless of how or where you generated them. It will use them as-is if
it can (the typical case), or process them if it needs to (the GN case).

This also changes the canonical compilation database from a symlink to a
a real file, which means we update the IDE's clangd settings when
changing targets instead. This allows clangd to automatically detect
when the database changes and restart itself, which it does not do for
symlinks. This means that in many cases, it will not be necessary to run
`pw ide sync` very frequently, since rebuilds that update the
compilation database will be automatically applied to clangd.

Original-Bug: b/272566137
Original-Bug: b/272566742
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/115445

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 62f10953b534851..9d63570df468cc2
Roller-URL: https://ci.chromium.org/b/8784182579663354145
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I855ae2f24f36e94cd605856b560fa882147159c4
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/138092
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>