[roll third_party/pigweed] pw_rpc: Add packet processors to ClientServerTestingContext
This CL adds an ability to customize packet processing in
ClientServerTestingContext (and its threaded variant).

This can be useful when the project has its own packet processing (e.g.
one that offloads all rpc handling to a custom thread pool) and wants to
simulate this behavior in tests as well.

This CL is a part of the chain and adds this support only to the basic
classes in pw::rpc::internal. It also adds a temporary constructor for
the backward compatibility that is removed as the last CL in the chain.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/117090

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: a1e8808c574a645..22e3d03a71c28dd
Roller-URL: https://ci.chromium.org/b/8779544959890846433
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I38f49d62d4f46fe705ccfdb5cb1ca0d680cbd38c
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/149830
Bot-Commit: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>
Commit-Queue: Pigweed Integration Roller <pigweed-integration-roller@pigweed.google.com.iam.gserviceaccount.com>