[roll third_party/pigweed] pw_chre: Add barebones CHRE
NOTE: This is a very work-in-progress implementation.

Add GN rules for the CHRE codebase into //third_party/CHRE.
This also adds a pw_chre_PLATFORM_BACKEND gn argument for the
platform backend for CHRE.

Add a pw_chre directory that includes all of the class and
function definitions required to compile CHRE. Some of this
connects to pigweed functionality, some of it is stubbed out.

//pw_chre:chre_simulator uses pw_system to run CHRE on
pigweed on the host, and it compiles in a number of CHRE's
example nanoapps. There is enough functionality that these
nanoapps are able to run.

Original-Bug: 294106526

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/162510

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: ef447ae6f95cab0..f210a064bf6d67f
Roller-URL: https://ci.chromium.org/b/8770150885417481937
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I89d476e84f1b5d0730d331078c2e2fce69a4e70c
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/170590
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>