[roll third_party/pigweed] pw_containers: VariableLengthEntryDeque C implementation
A VariableLengthEntryDeque is a double-ended queue buffer that stores
variable-length entries inline in a circular (ring) buffer. An entry's
length is stored as an LEB128 varint prefix.

This deque is a simplified, limited version of the
PrefixedEntryRingBuffer, and is implemented in C instead of C++. The C++
API will be added in a subsequent change.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/169731

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 5883b6332831e2e..d7606ceb99cc91c
Roller-URL: https://ci.chromium.org/b/8768845047394413473
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I8acb01c3afbb3378032f9f07f1368de695ea7232
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/173014
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>