[roll third_party/pigweed] pw_cli: Narrow logic around colors
We would always use colors if PW_USE_COLOR was set to a "true" value, or
if it was not set, or if we're connected to a terminal. Change that
logic to default to using colors if connected to a terminal, and
overriding that default if PW_USE_COLOR is set.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/173232

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 3a9727a30897a60..4a65df27bf9e41d
Roller-URL: https://ci.chromium.org/b/8768763302072639857
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: If29d9c2b40e041162e65b8b9da32c3f45a3386e9
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/173271
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>