[roll third_party/pigweed] pw_emu: core: Let the OS terminate foreground emulator processes
When the CLI runs the emulator in foreground there is no need to
explicitly terminate the emulator processes, the OS will do that for
us when the CLI process is terminated as long as the CLI runs attached
to a console.

Foreground CLI tests do not have a console attached (due to missing
pty support on Windows) and instead are using plain pipes. Because of
that logic was added to the core code to terminate the emulator
processes when it was running in foreground.

However, foreground tests were refactored to run the CLI as a session
leader and that allows clean termination for emulator processes
directly from the tests.

Thus it is no longer necessary to explicitly terminate emulator
processes in the core code when running in foreground.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/175638

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 34a7853d07c0014..36bca0bd5ec9216
Roller-URL: https://ci.chromium.org/b/8767056010297715905
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I7cf6e90bb808a2f8d36a3b9748b36e36f8f04961
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/176230
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>