[roll third_party/pigweed] pw_presubmit: Handle missing upstream better
If we try to find the commit at the remote tracking branch but are
unsuccessful, use 'HEAD~10'. This is excessive for individual CLs, but
not overly so, and it should be enough for most stacks of CLs.

Original-Bug: b/282808936
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/177038

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: 20847156966b050..dd93fa16b9de0a3
Roller-URL: https://ci.chromium.org/b/8765781233883776593
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I97bed7fbe8b9f7654dd201f42826c981d5f4dfe5
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/sample_project/+/178392
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>