[roll third_party/pigweed] pw_allocator: Make TrackingAllocator correct by construction
This CL removes the `Init` method from `TrackingAllocator` and
requires objects of that type to always be constructed with a valid
allocator. Previously, allowing the pointer to the wrapped allocator
to be null was believed to be needed to allow trackers around allocators
that might be optional depending on the platform. However, there is
already an allocator that can be used for this purpose: the
NullAllocator. As a result, the wrapped allocator can be a reference
instead of a pointer, and several conditional statement can be removed.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/193330

https://pigweed.googlesource.com/pigweed/pigweed
third_party/pigweed Rolled-Commits: e8c65f4cc0c1568..40fa6880873d75c
Roller-URL: https://ci.chromium.org/b/8755283815493821073
GitWatcher: ignore
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I07f501c7820a11ee447f2402a6ae5d27e190e5af
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/examples/+/193625
Bot-Commit: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Pigweed Roller <pigweed-roller@pigweed-service-accounts.iam.gserviceaccount.com>